+++ src/main/php/ValueReader.php
@@ -15,6 +15,10 @@ use stubbles\input\filter\range\DateRange;
 use stubbles\input\filter\range\DatespanRange;
 use stubbles\input\filter\range\StringLength;
 use stubbles\input\filter\range\NumberRange;
+use stubbles\peer\IpAddress;
+
+use function stubbles\input\predicate\isOneOf;
+use function stubbles\values\pattern;
 /**
  * Value object for request values to filter them or retrieve them after validation.
  *
@@ -428,11 +432,12 @@ class ValueReader implements valuereader\CommonValueReader
      */
     public function ifIsIpAddress()
     {
-        return $this->when(
-                \stubbles\predicate\IsIpAddress::instance(),
-                'INVALID_IP_ADDRESS',
-                []
-        );
+        if (IpAddress::isValid($this->param->value())) {
+            return $this->param->value();
+        }
+
+        $this->paramErrors->append($this->param->name(), 'INVALID_IP_ADDRESS');
+        return null;
     }
 
     /**
@@ -444,11 +449,34 @@ class ValueReader implements valuereader\CommonValueReader
      */
     public function ifIsOneOf(array $allowedValues)
     {
-        return $this->when(
-                new \stubbles\predicate\IsOneOf($allowedValues),
+        if (isOneOf($allowedValues)->test($this->param->value())) {
+            return $this->param->value();
+        }
+
+        $this->paramErrors->append(
+                $this->param->name(),
                 'FIELD_NO_SELECT',
                 ['ALLOWED' => join('|', $allowedValues)]
         );
+        return null;
+    }
+
+    /**
+     * returns value if it is matched by given regular expression
+     *
+     * @api
+     * @param   string  $regex  regular expression to apply
+     * @return  string
+     * @since   6.0.0
+     */
+    public function ifMatches($regex)
+    {
+        if (pattern($regex)->matches($this->param->value())) {
+            return $this->param->value();
+        }
+
+        $this->paramErrors->append($this->param->name(), 'FIELD_WRONG_VALUE');
+        return null;
     }
 
     /**
@@ -457,14 +485,11 @@ class ValueReader implements valuereader\CommonValueReader
      * @api
      * @param   string  $regex  regular expression to apply
      * @return  string
+     * @deprecated  since 6.0.0, use ifMatches() instead, will be removed with 7.0.0
      */
     public function ifSatisfiesRegex($regex)
     {
-        return $this->when(
-                new \stubbles\predicate\Regex($regex),
-                'FIELD_WRONG_VALUE',
-                []
-        );
+        return $this->ifMatches($regex);
     }
 
     /**
@@ -519,9 +544,9 @@ class ValueReader implements valuereader\CommonValueReader
      * If value does not satisfy the predicate return value will be null.
      *
      * @api
-     * @param   \stubbles\predicate\Predicate|callable  $predicate  predicate to use
-     * @param   string                                  $errorId    error id to be used in case validation fails
-     * @param   array                                   $details    optional  details for param error in case validation fails
+     * @param   \stubbles\input\predicate\Predicate|callable  $predicate  predicate to use
+     * @param   string                                        $errorId    error id to be used in case validation fails
+     * @param   array                                         $details    optional  details for param error in case validation fails
      * @return  string
      * @since   3.0.0
      */