+++ src/main/php/valuereader/CommonValueReader.php
@@ -203,10 +203,20 @@ interface CommonValueReader
     public function ifIsOneOf(array $allowedValues);
 
     /**
+     * returns value if it is matched by given regular expression
+     *
+     * @param   string  $regex  regular expression to apply
+     * @return  string
+     * @since   6.0.0
+     */
+    public function ifMatches($regex);
+
+    /**
      * returns value if it complies to a given regular expression, and null otherwise
      *
      * @param   string  $regex    regular expression to apply
      * @return  string
+     * @deprecated  since 6.0.0, use ifMatches() instead, will be removed with 7.0.0
      */
     public function ifSatisfiesRegex($regex);