+++ src/main/php/valuereader/MissingValueReader.php
@@ -300,15 +300,28 @@ class MissingValueReader implements CommonValueReader
     }
 
     /**
+     * returns value if it is matched by given regular expression
+     *
+     * @param   string  $regex  regular expression to apply
+     * @return  string
+     * @since   6.0.0
+     */
+    public function ifMatches($regex)
+    {
+        $this->reportError();
+        return null;
+    }
+
+    /**
      * returns value if it complies to a given regular expression, and null otherwise
      *
      * @param   string  $regex    regular expression to apply
      * @return  string
+     * @deprecated  since 6.0.0, use ifMatches() instead, will be removed with 7.0.0
      */
     public function ifSatisfiesRegex($regex)
     {
-        $this->reportError();
-        return null;
+        return $this->ifMatches($regex);
     }
 
     /**