+++ src/test/php/broker/param/MailParamBrokerTest.php
@@ -23,7 +23,6 @@ class MailParamBrokerTest extends MultipleSourceParamBrokerTest
      */
     public function setUp()
     {
-        $this->markTestIncomplete();
         $this->paramBroker = new MailParamBroker();
     }