+++ src/test/php/filter/MailFilterTest.php
@@ -35,7 +35,6 @@ class MailFilterTest extends FilterTest
      */
     public function setUp()
     {
-        $this->markTestIncomplete();
         $this->mailFilter = MailFilter::instance();
         parent::setUp();
     }